changed all frequency types to double

+++ src/FreqTable.java
@@ -1,21 +1,21 @@
 import java.util.Map;
 import java.util.HashMap;
 public class FreqTable {
-    private Map<Character, Integer> freqTable;
+    private Map<Character, Double> freqTable;
     /**
      * A simply wrapper class for a Hashmap 
      * representing the frequencies of characters in our alphabet
      */
     public FreqTable() {
-        this.freqTable = new HashMap<Character, Integer>();
+        this.freqTable = new HashMap<Character, Double>();
     }
 
     /**
      * @return A deep copy of the frequency table
      */
-    public Map<Character, Integer> getFreqTable() {
-        Map<Character, Integer> freqTableClone = new HashMap<Character, Integer>();
-        for(Map.Entry<Character, Integer> entry : this.freqTable.entrySet()) {
+    public Map<Character, Double> getFreqTable() {
+        Map<Character, Double> freqTableClone = new HashMap<Character, Double>();
+        for(Map.Entry<Character, Double> entry : this.freqTable.entrySet()) {
             freqTableClone.put(entry.getKey(), entry.getValue());
         }
 
@@ -28,7 +28,7 @@ public class FreqTable {
      * @param freq The frequency of this character
      * @throws IllegalArgumentException if the character is already in the Hashmap
      */
-    public void addCharacter(char c, int freq) throws IllegalArgumentException {
+    public void addCharacter(char c, double freq) throws IllegalArgumentException {
         if(this.freqTable.containsKey(c)) {
             throw new IllegalArgumentException("Character is already present in Frency Table");
         }
@@ -41,7 +41,7 @@ public class FreqTable {
      * @param freq The new frequency of the character
      * @throws IllegalArgumentException if the character does not exist
      */
-    public void updateFreq(char c, int freq) throws IllegalArgumentException {
+    public void updateFreq(char c, double freq) throws IllegalArgumentException {
         if(!this.freqTable.containsKey(c)) {
             throw new IllegalArgumentException("Character is not present in Frequency Table");
         }
@@ -53,7 +53,7 @@ public class FreqTable {
      * E.g. if the character 'c' appears with frequency 3, displays "c: 3"
      */
     public void print() {
-        for(Map.Entry<Character, Integer> entry : this.freqTable.entrySet()) {
+        for(Map.Entry<Character, Double> entry : this.freqTable.entrySet()) {
             System.out.println(entry.getKey() + ": " + entry.getValue());
         }
     }
