+++ src/HInternalNode.java
@@ -1,5 +1,5 @@
 public class HInternalNode extends HNode {
-    public HInternalNode(int _freq){
+    public HInternalNode(double _freq){
         this.setFreq(_freq);
         this.setSymbol('\0'); //Internal Nodes have NULL symbol
     }
