+++ src/HLeafNode.java
@@ -1,6 +1,6 @@
 public class HLeafNode extends HNode {
     
-    public HLeafNode(char _symbol, int _freq) {
+    public HLeafNode(char _symbol, double _freq) {
         this.setSymbol(_symbol);
         this.setFreq(_freq);
     }
